requirements.txt
Flask==3.0.0
PyYAML==6.0.1
python-dotenv==1.0.0
